+++ assets/src/edit-story/components/canvas/displayElement.js
@@ -30,8 +30,8 @@ import { useRef } from '@wordpress/element';
 import { getDefinitionForType } from '../../elements';
 import { elementWithPosition, elementWithSize, elementWithRotation } from '../../elements/shared';
 import StoryPropTypes from '../../types';
+import { useTransformHandler } from '../transform';
 import { useUnits } from '../../units';
-import useTransformHandler from './useTransformHandler';
 
 const Wrapper = styled.div`
 	${ elementWithPosition }
