+++ assets/src/edit-story/components/canvas/index.js
@@ -16,4 +16,3 @@
 
 export { default } from './canvas';
 export { default as useCanvas } from './useCanvas';
-export { default as useTransformHandler } from './useTransformHandler';
