+++ assets/src/edit-story/components/canvas/singleSelectionMovable.js
@@ -32,6 +32,7 @@ import Movable from '../movable';
 import calculateFitTextFontSize from '../../utils/calculateFitTextFontSize';
 import objectWithout from '../../utils/objectWithout';
 import getAdjustedElementDimensions from '../../utils/getAdjustedElementDimensions';
+import { useTransform } from '../transform';
 import { useUnits } from '../../units';
 import { MIN_FONT_SIZE, MAX_FONT_SIZE } from '../../constants';
 import { getDefinitionForType } from '../../elements';
@@ -49,8 +50,9 @@ function SingleSelectionMovable( {
 	const [ isResizingFromCorner, setIsResizingFromCorner ] = useState( true );
 
 	const { actions: { updateSelectedElements }, state: { currentPage } } = useStory();
-	const { actions: { pushTransform }, state: { pageSize: { width: canvasWidth, height: canvasHeight }, nodesById } } = useCanvas();
+	const { state: { pageSize: { width: canvasWidth, height: canvasHeight }, nodesById } } = useCanvas();
 	const { actions: { getBox, dataToEditorY, editorToDataX, editorToDataY } } = useUnits();
+	const { actions: { pushTransform } } = useTransform();
 
 	const minMaxFontSize = {
 		minFontSize: dataToEditorY( MIN_FONT_SIZE ),
