+++ includes/templates/single-web-story.php
@@ -46,11 +46,11 @@ use \Google\Web_Stories\Media;
 		.page-safe-area {
 			position: absolute;
 			overflow: hidden;
-	    margin: auto;
-	    top: 0;
-	    bottom: 0;
-	    left: 0;
-	    right: 0;
+		margin: auto;
+		top: 0;
+		bottom: 0;
+		left: 0;
+		right: 0;
 		}
 
 		.wrapper {
@@ -64,7 +64,7 @@ use \Google\Web_Stories\Media;
 			left: 0;
 			right: 0;
 			bottom: 0;
-	    margin: 0;
+		margin: 0;
 		}
 	</style>
